requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
mysql-connector-python>=8.1.0
psycopg2-binary>=2.9.7
pymongo>=4.5.0
pydantic>=2.0.0
typing-extensions>=4.7.0
